Add --verbose option
C++ standard library does not provide command line parsing and getopt is
not available on Windows, so here is a
Yet Another Home Grown Incomplete And Incompatible Command Line Parser™.

May be added a --quite option and make the verbose mode a bit more
verbose, but this should do for now.

Fixes #100.