[ots] Always check the dynamic_cast
Previously some places relied on prior checks for the table presence,
but this can’t be relied on now since the table might be a path thru
table that we didn’t parse which would pass the first check but the
dynamic_cast would still result in a NULL pointer. Now check the table
when needed and drop other superfluous checks.