[no-rtti] Eliminate use of dynamic_cast so OTS can be built with -fno-rtti
Instead of relying on C++ RTTI, this adds an m_type field to Table, and
each specific subclass constructor initializes this to the tag that
corresponds to its type. This is used to implement a
Font::GetTypedTable() method that checks whether the returned Table does
in fact have the expected type (i.e. it wasn't loaded as unparsed data
using TablePassthru), in which case it can safely be down-cast with
static_cast<> by the caller.

Fixes https://github.com/khaledhosny/ots/issues/129.