[graphite] Fix binary-search header fields rather than dropping table
The correct values for these fields can be computed from the number of entries
in the list to be searched, so we can simply overwrite them with the proper
values rather than rejecting tables where they appear wrong.

(Existing fonts in circulation often have bad values in these fields; in
practice, it doesn't matter because the current engine does not rely on
them, and therefore the errors have not been noticed. But in case any future
engine does try to use them, we should fix them up here.)