[graphite] Fix invalid lbGID instead of dropping table
This field is only used by the Graphite justifier, which no client applications are
currently (known to be) using. The value in Awami (at least) is incorrect due to a
bug in the subsetter being used to generate the font served on the Graphite demo page,
but as it is unused, the error is harmless and we shouldn't reject the font.

Setting to zero should be benign; at worst, I guess a client that did attempt to use the
justifier might see spurious .notdef glyphs, but there should be no security issues
as a result.