[STAT] Add null-check to silence coverity warning
Coverity is worried that `name` could be nullptr here (as `GetTypedTable` can return null). This won't actually happen, because `name` is a required table and so OTS will reject the font before it even attempts to parse `STAT` if it is missing; but I suggest we add a null-check here rather than relying on the order of table processing.