[tests] Remove duplicate font file
The same fie exists under tests/fonts/bad and it breaks oss-fuzz when it
tries to zip the font file. All font files under tests/ are used to seed the
fizzer, files under tests/fuzzing are fuzzed files that were generated
by oss-fuzz and uncovered some issue with the code.