meson.build: use dependencies, remove third_party/
When downloading releases of the ots, the .tar.gz source archives are
not including the third_party/ code. This makes the releases of the
package not buildable from source.

This commit removes the hardcoding of the third_party libraries, and
use the meson dependency mechanism to find the correct libraries on
any host.

While this move is bold and may displease upstream, it is necessary
in order to provide this tool in distributions that do not encourage
the embedding of libraries inside of each other, but promotes the
shared library approach.

If this is not acceptable as-is, we should find a way in the meson.build
to try to use the system dependencies, and if they are not found
fall-back to the copied third_party sources.

This change removes the dirty hacks in meson.build aswell, because they
are figured out by the libraries themselves.