[variations] Synthesize an empty 'gvar' if missing.
If there is an 'fvar' table that defines variation axes, some rasterizers (e.g. Core Text)
assume this means 'gvar' will be present, and fail if it is missing. To work around this,
add an empty 'gvar' to such fonts.

This behavior is controlled by the --synthesize-gvar configuration option, so it can be
disabled if not required in a given build.