[VDMX] Correct VDMX.numRecs field if necessary, to ensure enough records are retained.
Without this, sanitizing the ARIALBD.ttf file referenced at https://bugzilla.mozilla.org/show_bug.cgi?id=1761233#c25
(apparently from the old MS Core Fonts, but may have been modified?) generates a bad result, because numRecs in the font
is 1, but in fact there are 4 records, and offsets pointing to them, so we need them all to be preserved.

(Also consistently prefer Drop() to Error() when parsing VDMX, as the font should still be functional without it.)

Added a test font created by pasting a bad (but fixable) VDMX table into a copy of Ahem.ttf (hacked to make the
glyph count match), as I'm not comfortable directly using the proprietary font as testcase. Prior to the patch here,
this font fails in ots-idempotent because the first sanitization creates a broken VDMX, which then fails to parse
on the second attempt.