[maxp] Attempt to improve fixup of bad 'maxp' version numbers.
If the version number is not explicitly v0.5 (0x00005000), try to read
it as v1.0. (The version field will be corrected to one of the valid
versions during serialization.)

This handles cases where the table was intended to be v1.0 (for a font
with TrueType data), but the version field was set incorrectly (e.g. to
0x00000100, as seen in https://github.com/mozilla/pdf.js/issues/16839).

Also make the max_zones fixup more general, as the example font from the
above issue has max_zones=512(!). Correcting it to 2 should be harmless.