[Graphite] Add assertions in OpenTypeGLAT_*::Serialize methods.
It turns out that we don't handle "dropping" the Glat table correctly,
because the m_shouldSerialize flag gets cleared on the specific
implementation subclass, but then the OTS driver code checks the
flag on the OpenTypeGLAT wrapper, and may proceed to call Serialize
for a table that should've been dropped.

Here, we add assertions to the Serialize method in the implementation
subclasses, to check that it is not called if the table is supposed
to be excluded.

(It turns out that some existing tests can hit these assertions,
in addition to the newly-added testcase.)