New overload OTSContext::Process to lower memory use processing WOFF2
NOTE: This patch is intrusive. It's not necessarily expected that this
patch can land in its current state, rather it is provided to spur
discussion.

We are Chromium integrators and we've seen that processing some WOFF2
fonts can take up a lot of memory. This patch adds a new overload to
OTSContext::Process in order to make it possible to use less memory
while processing WOFF2 fonts. It does so via the following changes:

- Makes ots::ExpandingMemoryStream part of public ots API.
- Makes ots::ExpandingMemoryStream handle move-construct and
  move-assignment.
- Adds an overload of ots::OTSContext::Process that takes an rvalue ref
  to ots::ExpandingMemoryStream as input buffer.
- Makes ots::OTSContext::Process return the output buffer instead of
  writing to a separate output buffer.