Introduce new TABLE_ACTION_SANITIZE_SOFT action.
This action (not used as the default for any table tags) performs the same
table-specific Parse() as TABLE_ACTION_SANITIZE, but if parsing fails, it will
drop the table but _not_ return a failure status to the overall processing.

This can be used by a client that wishes to continue processing and using a
resource even if certain tables may fail sanitization and be discarded.

It differs from SANITIZE in that failure to parse does not completely stop
processing of the resource and return an overall failure.

It differs from PASS_THRU in that the table is subject to sanitization,
reporting any errors found, and will be dropped if it fails, rather than
silently accepted.

It differs from DROP in that the table is only dropped if (unrecoverable)
errors are found by Parse(); if the table is valid, it will be retained.